Move interop classes to ChartJs.Blazor.Interop
