+++ samples/ClientSide/ChartJs.Blazor.Sample.ClientSide/Pages/Test.razor
@@ -10,6 +10,7 @@
 @using ChartJs.Blazor.ChartJS.Common.Wrappers
 @using ChartJs.Blazor.ChartJS.Common.Axes.Ticks
 @using ChartJs.Blazor.ChartJS.Common.Handlers
+@using ChartJs.Blazor.Interop
 @using System.Text.Json
 
 <ChartJsBarChart @ref="barChartJs" Config="@_config" Height="250"/>
