+++ samples/ServerSide/ChartJs.Blazor.Sample.ServerSide/Pages/Test.razor
@@ -13,6 +13,7 @@
 @using ChartJs.Blazor.ChartJS.Common.Wrappers
 @using ChartJs.Blazor.ChartJS.Common.Axes.Ticks
 @using ChartJs.Blazor.ChartJS.Common.Handlers
+@using ChartJs.Blazor.Interop
 @using System.Text.Json
 
 <h1>Pie Chart with legend callbacks</h1>
