+++ samples/Shared/ChartJs.Blazor.Sample.Shared/Components/Callbacks/ClickAndHoverHandlerComponent.razor
@@ -1,6 +1,7 @@
 @using System.Globalization
 @using ChartJs.Blazor.ChartJS.Common
 @using ChartJs.Blazor.ChartJS.Common.Handlers
+@using ChartJs.Blazor.Interop
 @using ChartJs.Blazor.Charts
 @using ChartJs.Blazor.ChartJS.PieChart
 @using ChartJs.Blazor.ChartJS.Common.Properties
