+++ samples/Shared/ChartJs.Blazor.Sample.Shared/Components/Callbacks/LegendCallbacksComponent.razor
@@ -4,6 +4,7 @@
 @using ChartJs.Blazor.ChartJS.Common
 @using ChartJs.Blazor.ChartJS.Common.Properties
 @using ChartJs.Blazor.ChartJS.Common.Handlers
+@using ChartJs.Blazor.Interop
 @using ChartJs.Blazor.Util
 @using System.Text.Json
