+++ samples/Shared/ChartJs.Blazor.Sample.Shared/Components/Fun/SineLineChartComponent.razor
@@ -2,6 +2,7 @@
 @using ChartJs.Blazor.Charts
 @using ChartJs.Blazor.ChartJS.Common
 @using ChartJs.Blazor.ChartJS.Common.Handlers
+@using ChartJs.Blazor.Interop
 @using ChartJs.Blazor.ChartJS.Common.Axes
 @using ChartJs.Blazor.ChartJS.Common.Axes.Ticks
 @using ChartJs.Blazor.ChartJS.Common.Properties
