+++ samples/Shared/ChartJs.Blazor.Sample.Shared/Components/LineChart/LinearLineChartComponent.razor
@@ -6,6 +6,7 @@
 @using ChartJs.Blazor.ChartJS.Common.Axes
 @using ChartJs.Blazor.ChartJS.Common.Axes.Ticks
 @using ChartJs.Blazor.ChartJS.Common.Handlers
+@using ChartJs.Blazor.Interop
 @using ChartJs.Blazor.Sample.Shared.Services
 @using ChartJs.Blazor.Util
 @using System.Text.Json
