+++ samples/Shared/ChartJs.Blazor.Sample.Shared/Components/LineChart/TimeLineChartComponent.razor
@@ -8,6 +8,7 @@
 @using ChartJs.Blazor.ChartJS.Common.Time
 @using ChartJs.Blazor.ChartJS.LineChart
 @using ChartJs.Blazor.Sample.Shared.Services
+@using ChartJs.Blazor.Interop
 @using ChartJs.Blazor.Util
 @inject IJSRuntime JsRuntime
