+++ src/ChartJs.Blazor/ChartJS/Common/BaseConfigOptions.cs
@@ -1,6 +1,7 @@
 ﻿using ChartJs.Blazor.ChartJS.Common.Enums;
 using ChartJs.Blazor.ChartJS.Common.Handlers;
 using ChartJs.Blazor.ChartJS.Common.Properties;
+using ChartJs.Blazor.Interop;
 
 namespace ChartJs.Blazor.ChartJS.Common
 {
