+++ src/ChartJs.Blazor/ChartJS/Common/Properties/Legend.cs
@@ -1,5 +1,6 @@
 ﻿using ChartJs.Blazor.ChartJS.Common.Enums;
 using ChartJs.Blazor.ChartJS.Common.Handlers;
+using ChartJs.Blazor.Interop;
 
 namespace ChartJs.Blazor.ChartJS.Common.Properties
 {
